\name{apeTests-package}
\alias{apeTests-package}
\alias{apeTests}
\docType{package}
\title{
Tests for Package ape
}
\description{
  \pkg{apeTests} provides tests for ape.

  The complete list of functions can be displayed with
  \code{library(help = apeTests)}.

  More information on \pkg{ape} can be found at \url{https://emmanuelparadis.github.io/}.
}

\author{
  Emmanuel Paradis, Martin Smith, Thomas Guillerme

  Maintainer: Emmanuel Paradis <Emmanuel.Paradis@ird.fr>
}
\references{
  Paradis, E. (2012) \emph{Analysis of Phylogenetics and Evolution with
    R (Second Edition).} New York: Springer.

  Paradis, E., Claude, J. and Strimmer, K. (2004) APE: analyses of
  phylogenetics and evolution in R language. \emph{Bioinformatics},
  \bold{20}, 289--290.

  Popescu, A.-A., Huber, K. T. and Paradis, E. (2012) ape 3.0: new tools
  for distance based phylogenetics and evolutionary analysis in
  R. \emph{Bioinformatics}, \bold{28}, 1536--1537.

  Paradis, E. and Schliep, K. (2019) ape 5.0: an environment for modern
  phylogenetics and evolutionary analyses in R. \emph{Bioinformatics},
  \bold{35}, 526--528.
}
\keyword{package}
